+++ src/classes/incubation/distros/buster.ts
@@ -12,11 +12,11 @@ import shx from 'shelljs'
 import yaml from 'js-yaml'
 import path from 'node:path'
 
-import {IInstaller, IRemix, IDistro} from '../../../interfaces/index.js'
+import {IInstaller, IRemix, IDistro} from '../../../interfaces/index'
 
-import Fisherman from '../fisherman.js'
+import Fisherman from '../fisherman'
 
-import {exec} from '../../../lib/utils.js'
+import {exec} from '../../../lib/utils'
 
 /**
  *
@@ -36,13 +36,15 @@ export class Buster {
 
   theme: string
 
+  isClone: boolean
+
   /**
    * @param remix
    * @param distro
    * @param displaymanager
    * @param verbose
    */
-  constructor(installer: IInstaller, remix: IRemix, distro: IDistro, user_opt: string, release = false, theme = 'eggs', verbose = false) {
+  constructor(installer: IInstaller, remix: IRemix, distro: IDistro, user_opt: string, release = false, theme = 'eggs', isClone = false, verbose = false) {
     this.installer = installer
 
     this.remix = remix
@@ -51,6 +53,7 @@ export class Buster {
     this.verbose = verbose
     this.release = release
     this.theme = theme
+    this.isClone = isClone
   }
 
   /**
@@ -59,7 +62,7 @@ export class Buster {
   async create() {
     const fisherman = new Fisherman(this.distro, this.installer, this.verbose)
 
-    await fisherman.settings(this.theme)
+    await fisherman.settings(this.theme, this.isClone)
 
     await fisherman.buildModule('partition', this.theme)
     await fisherman.buildModule('mount')