+++ src/components/location.tsx
@@ -4,12 +4,12 @@
 
 import React, { useState } from 'react'
 import { render, Text, Box, Newline } from 'ink'
-import Title from './elements/title.js'
-import Steps from './elements/steps.js'
+import Title from './elements/title'
+import Steps from './elements/steps'
 
 import yaml from 'js-yaml'
 import fs from 'fs'
-import { ISettings, IBranding } from '../interfaces/index.js'
+import { ISettings, IBranding } from '../interfaces/index'
 
 
 type LocationProps = {