+++ src/krill/modules/grubcfg.ts
@@ -7,8 +7,8 @@
  * https://stackoverflow.com/questions/23876782/how-do-i-split-a-typescript-class-into-multiple-files
  */
 
-import Sequence from '../krill-sequence.js'
-import Utils from '../../classes/utils.js'
+import Sequence from '../krill-sequence'
+import Utils from '../../classes/utils'
 import fs from 'fs'
 
 /**