+++ src/krill/modules/network-cfg.ts
@@ -7,9 +7,9 @@
  * https://stackoverflow.com/questions/23876782/how-do-i-split-a-typescript-class-into-multiple-files
  */
 
-import Sequence from '../krill-sequence.js'
-import Utils from '../../classes/utils.js'
-import Pacman from '../../classes/pacman.js'
+import Sequence from '../krill-sequence'
+import Utils from '../../classes/utils'
+import Pacman from '../../classes/pacman'
 
 /**
 /**