+++ changelog.md
@@ -21,6 +21,10 @@ You can follow the project also consulting the [commit history](https://github.c
 ## Changelog
 Versions are listed on reverse order, the first is the last one. Old versions are moved to [versions](https://sourceforge.net/projects/penguins-eggs/files/packages-deb/versions/). 
 
+
+### eggs-9.3.30
+* produce --clone: now calamares - when configured for clone - no longer asks for user configuration and previous users are used. You can also install with krill; the advantage is that it will be faster and will respect the autologin configuration, which is reset by calamares to the default.
+
 ### eggs-9.3.29
 * fix: produce --clone unmount all binded mounts;
 * fix: in case of errors, now produces shows the error and pauses before to end.